server/requirements.txt
fastapi 
sqlalchemy 
psycopg2-binary 
uvicorn 
requests
python-dotenv
openai
langchain
langgraph
langchain-openai
uvicorn[standard]
pgvector